requirements.txt
browser-use==0.1.40   
fastapi==0.115.12
langchain-openai==0.3.1
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.0
nest-asyncio==1.6.0
html2text==2024.2.26
pillow==11.1.0